+++ arch/lib/registers.h
@@ -1,6 +1,8 @@
 #ifndef ARCH_LIB_REGISTERS_H
 #define ARCH_LIB_REGISTERS_H
 #include <stdint.h>
+#include <stdlib.h>
+#include "log.h"
 
 #define ERROR_WRONG_REGISTER_SELECTED 1
 #define ERROR_WRONG_REGISTER_VALUE 2
@@ -15,17 +17,17 @@ typedef enum {
 } sc_flag_t;
 
 typedef struct {
-    int16_t accum;
-    int8_t count;
-    int8_t flag;
+    int accum;
+    int count;
+    int flag;
 } sc_reg_t;
 sc_reg_t sc_register;
 
 int sc_regInit();
-int sc_regSet(int8_t register_id, int8_t value);
-int sc_regGet(int8_t register_id, int8_t * value);
+int sc_regSet(int register_id, int value);
+int sc_regGet(int register_id, int * value);
 int sc_accumGet(int * value);
 int sc_accumSet(int value);
-int sc_instGet(int8_t * value);
-int sc_instSet(int8_t value);
+int sc_instGet(int * value);
+int sc_instSet(int value);
 #endif
